# src/server/middleware/errorHandler/validationErrorCode.ts
export enum ValidationErrorCode {
    RequiredField = 1000,
    IncorrectFormat = 1001,
    InvalidAddress = 1002,
    AddressNotSupported = 1003,
    ValueOutOfRange = 1004,
    InvalidECDSAOrHash = 1005,
    UnsupportedOption = 1006,
}